[ref] Now use analyze(source) -> expressions; eval(expression) -> obj

+++ src/env.c
@@ -27,17 +27,30 @@ void set_env(lisp_env *env, const char *name, lisp_obj *value)
     assert(name != NULL);
     assert(value != NULL);
 
+    /* Hold ref for this obj if not owned by someone else */
+    if (value->env == NULL){
+        value->env = (void *) env;
+    }
+
+    /* If an object already has this name in this env, drop ref and detroy, 
+       then replace by new object */
     for (size_t i=0; i<env->used; i++){
         if (strcmp(env->names[i], name) == 0){
-            destroy_obj(env->values[i]);
+            if (env->values[i]->env == env){
+                env->values[i]->env = NULL;
+                destroy_obj(env->values[i]);
+            }
             env->values[i] = value;
             return;
         }
     }
 
+    /* If not found, and there are subsequent env buckets set in following */
     if (env->next){
         set_env(env->next, name, value);
     } 
+
+    /* Otherwise insert here */
     else {
         if (env->used == LISP_ENV_SIZE){
             env->next = create_env(env->parent);
@@ -61,7 +74,10 @@ lisp_env *destroy_env(lisp_env *env)
     lisp_env *parent = env->parent;
     for (size_t i=0; i<env->used; i++){
         free(env->names[i]);
-        destroy_obj(env->values[i]);
+        if (env->values[i]->env == env){
+            env->values[i]->env = NULL;
+            destroy_obj(env->values[i]);
+        }
     }
     free(env);
     return parent;
